+++ torchbenchmark/models/timm_vision_transformer/__init__.py
@@ -55,7 +55,7 @@ class Model(BenchmarkModel):
 
 	def _step_eval(self):
 		nvtx.range_push('eval')
-		output = self.model(self.cfg.infer_example_inputs)
+		output = self.model(self.cfg.infer_example_inputs.half())
 		nvtx.range_pop()
 
 	def get_module(self):
@@ -96,7 +96,7 @@ class Model(BenchmarkModel):
 		self.model.eval()
 		torch.backends.cudnn.benchmark = True
 		with torch.no_grad():
-			# self.model = self.model.half()
+			self.model = self.model.half()
 			graphs=True
 			if graphs:
 				niter = 8
